Controllerのメソッドをasync/awaitで書き直した
1. Controllerのメソッド変更
2. Controllerのコンストラクタに渡す不要な引数を削除
2. bin、libディレクトリ作成
