+++ 04.class/bin/memo.js
@@ -1,4 +1,4 @@
-const Command = require('./command')
+const Command = require('../lib/command')
 
 const command = new Command(process.argv[2])
 command.run()
