--- 04.class/drop.js
@@ -1,7 +1,0 @@
-const sqlite = require('sqlite3').verbose()
-const db = new sqlite.Database('memos.sqlite')
-
-db.serialize(() => {
-  db.run('DROP TABLE memos')
-})
-db.close()
